Long awaited Ruffle update
Hey, so uh, I'm back working on this project. It's been a while, due to a multitude of reasons mainly due to other various hobbies, the summer, and, well, not as much as a need. Let me put it like this. When I made this website, EVERYONE was playing games during class, we were online it made sense. It also came mere months after Flash dead, so a mixture of Nostalgia and need were there a Viola, the website was born. However, now, we're not online really much at all. And the ease of access to old flash content is easier than ever. With Ruffle officially being an extension for chrome now. But, hey. This was a fun project. And I can't wait to keep working on it even though the need is really there. Alright, Adios everyone. Peace. Bye.